Renammed code
Just renammed some util files because 'Util' is not the best name.
